refs.c: add a function to append a reflog entry to a fd
Move code to create the string for a ref and write it to a file
descriptor from log_ref_write and add it into a new dedicated
function log_ref_write_fd.

For now the new function is only used from log_ref_write, but later
on we will call this function from reflog transactions too.  That
means that we will end up with only a single place, where we write
a reflog entry to a file instead of the current two places
(log_ref_write and builtin/reflog.c).

Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>
Signed-off-by: Stefan Beller <sbeller@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>